move app-related constants to own file
